requirements.txt
Flask==1.1.2
Flask-Restful==0.3.8
requests==2.25.1
numpy==1.20.3
python-dotenv==0.17.1
ecdsa==0.14.1
cffi>=1.15.0
cryptography>=37.0.4
pycparser>=2.21
python-socks>=2.1.1
secp256k1>=0.14.0
websocket-client>=1.3.3
pytest>=7.2.0
pytest-cov[all]